requirements.txt
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
Flask==2.0.1
google-api-core==1.31.1
google-api-python-client==2.15.0
google-auth==1.34.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==3.2
importlib-metadata==4.6.4
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==21.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
PySide2==5.15.2
PySimpleGUIQt==0.35.0
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
rsa==4.7.2
shiboken2==5.15.2
six==1.16.0
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
Werkzeug==2.0.1
zipp==3.5.0
